Fixed bug in clone (git cannot work with / in path very wel, replaced it with /)

+++ GitUI/FormDeleteBranch.cs
@@ -26,7 +26,10 @@ namespace GitUI
         {
             try
             {
-                MessageBox.Show("Command executed \n" + GitCommands.GitCommands.DeleteBranch(Branches.Text), "Delete branch");
+                if (MessageBox.Show("Are you sure you want to delete this branch?\nDeleting a branch can cause commits to be deleted too!", "Delete branch", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                {
+                    MessageBox.Show("Command executed \n" + GitCommands.GitCommands.DeleteBranch(Branches.Text), "Delete branch");
+                }
             }
             catch
             {
